Better handle dirty git repos used as infra template
Doesn't feel like this totally belongs in
`compute_app_includes_excludes`, but for potentially gitignored files
it's a little harder to generate a complete list of files to pass in
ahead of time (unlike untracked files), so individual calls to
`is_path_ignored()` require the git repo. Might be able refactor things
for better separation with some more effort.

Also have `InfraTemplate` require the directory it's passed is a git
working directory, since that seems to be the actual current
requirement.