Get tool running via docker (#2)
Fixup docker things so it will actually work.

Automate the docker volume mounting in `bin/docker-wrapper`. This requires
parsing out some of the arguments to be passed to the platform executable itself
and transforming them. That arg processing is brittle at the moment, but trying
to keep things simple for now. Pass all options/named args first, with the
project directory as the last argument. If/when there are commands that don't
take a project directory, we'll have to revisit the setup.

Additionally provide an entry script, in `bin/docker-entry`, that will run the
tooling as the user that owns the project directory to mitigate file ownership
issues (particularly on Linux hosts).

If manually mounting volumes/running docker, be sure to mount the project
directory as `/project-directory`, if you want `bin/docker-entry` to use the
right permissions.