Reorganize the source code
Move things under a `nava.platform` package, with `nava.platform.cli`
for the bits of the actual tool and other things in
`nava.platform.utils` or in `nava.platform` itself for now. Some of
these are really specific to the platform CLI tool itself, but create a
little separation to see how it feels for a while.