Support installing apps with the same name as an existing directory
Do this by updating the exclude logic to add the specific
files/directories it is handling to the list rather than just their
name (which causes copier to ignore everything that might have the same
name).

Notably this applied to the `app/` example app that `template-infra` CI
wants to install as its app for basic validation of itself. We'll
probably rename `app/` anyway, but this update should generally be more
correct.