fail vs fail_with_usage
Click's (and so Typer's) `fail` raises a `UsageError` exception[1],
which by default will print the relevant command's help message[2]. Also
sets the exit code to 2. This can be helpful in some error scenarios,
like the user actually called the tool wrong or is something where the
command help provides more context.

But often the command's help page won't be super useful, and we just
want to print a message and exit, so have `fail` do that, with
`fail_with_usage` for Click's behavior.

[1] https://github.com/pallets/click/blob/fde47b4b4f978f179b9dff34583cb2b99021f482/src/click/core.py#L691
[2] https://github.com/pallets/click/blob/fde47b4b4f978f179b9dff34583cb2b99021f482/src/click/exceptions.py#L72