Refactor bin/docker-wrapper
Catch it up on some breaking changes and generally try to make it less
brittle.

Mount a host log directory for the tool inside docker to write to.

Mount the host gitconfig (if it exists), so git commits work inside
docker.

Try to mount any argument that looks like a host path into the docker
environment, with the same path. Dropping the special handling for the
"project dir" location.

Remove the setuid bits from `bin/docker-entry`. Using docker-wrapper is
pretty much required, so avoid extra work.