Add `--force` option to update commands
This effectively does a `cp -rf` of the template onto the project.
Unlike Copier's `recopy`, this allows you to change the
`--template-uri`/`_src_path` value.

This is not expected to be a common thing, but sometimes, particularly
when iterating on larger template changes/moving files around, Copier
can not detect a change you'd expect or ignores the file because the
project has deleted it at some point in the past (which is expected
behavior). This `--force` allows you to apply the updated template in a
brute force way. You can then pick out the pieces you want via the git
diff.

You probably want to combine this with `--no-commit`, since by default
the update commands will commit *all* the changes.