infra update-base: Use the template version that was updated to when regenerating network config
If no `--version` value is given on the CLI, `version` will be `None`,
in which case `_update_network_config()` will fallback to
`project.template_version` which requires the base and app templates all
be in sync version-wise, but we _just_ updated the base one before the
app ones, so that will never be the case.

Instead of passing through a potentially empty `version` and triggering
that behavior, use the version that the base template is now on.